Revert "checkoutUI update and some type changes"
This reverts commit 66e0c1516a5cf9ec868c859c68b7605a1628f4dc.

+++ src/objects/Discount.java
@@ -9,8 +9,8 @@ public class Discount {
      * @param member; an example of a trait that provides certain discount
      * @return 0.1 due to pre-set discount for parameter
      */
-    public static float getDiscount(Membership member){
-        return 0.1f; //10% discount for member
+    public static double getDiscount(Membership member){
+        return 0.1; //10% discount for member
     }
     //extendable - just overload method for more object types and corr. discount
 }
